Feature/pwa

+++ src/dashboard/assets/js/havenotAccess.js
@@ -1,17 +1,31 @@
-var BASE_URL = "https://e0c1beaf270b.ngrok.io/";
+var BASE_URL = "http://127.0.0.1:8000/";
 
-window.onload = function(){
+function verification_have_not_access(){
+    // console.log('check')
     if(localStorage.getItem("access")){
-        $.post(BASE_URL+'auth/login/verify/',{"token":localStorage.getItem("access")},function(data,status){            
-        }).fail(function(data){
-            localStorage.clear();
-            window.location.href = "../examples/login.html"
-        }).always(function(){
+        let timeNow = new Date().getTime()
+        lastVerfiyCall = localStorage.getItem("lastVerfiyCall")?new Date(localStorage.getItem("lastVerfiyCall")).getTime():new Date(2000,10,30).getTime()
+        // console.log((timeNow - lastVerfiyCall ) / 1000 )
+        
+        if((timeNow - lastVerfiyCall ) / 1000 > 900){
+            $.post(BASE_URL+'auth/login/verify/',{"token":localStorage.getItem("access")},function(data,status){            
+                console.log('have not access')
+                localStorage.setItem('lastVerfiyCall',new Date().toString()) 
+            }).fail(function(data){
+                localStorage.clear();
+                window.location.href = "../examples/login.html"
+                localStorage.removeItem("lastVerfiyCall")
+            }).always(function(){
+                
+            })
+        }else{
             
-        })
+        }
     }else{
         localStorage.clear();
         window.location.href = "../examples/login.html"
     }
      
 }
+
+window.addEventListener('load', verification_have_not_access)
